pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "autoantibody"
version = "0.0.1"
description = "Autoantibody analysis tools"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Bryan Briney", email = "briney@scripps.edu" },
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "biopython>=1.80",
    "numpy>=1.24",
    "typer>=0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
    "mypy>=1.0",
    "pre-commit>=3.0",
    "types-PyYAML>=6.0",
    "scipy>=1.10",
]
benchmark = [
    "pandas>=2.0",
    "pyarrow>=14.0",
    "requests>=2.28",
    "matplotlib>=3.7",
]

[project.urls]
Repository = "https://github.com/briney/autoantibody"

# --- Tool Configuration ---

[tool.ruff]
line-length = 100

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
]

[tool.ruff.lint.per-file-ignores]
# Pydantic models need runtime access to types like datetime and Path
"src/autoantibody/models.py" = ["TCH003"]
# container.py and run_tool.py use Path at runtime
"src/autoantibody/container.py" = ["TCH003"]
"src/autoantibody/run_tool.py" = ["B008", "TCH003"]
# Test files use Path at runtime in fixtures/signatures
"tests/*.py" = ["TCH003"]
# Typer uses Option() in function defaults by design; scripts use Path at runtime
"scripts/*.py" = ["B008", "TCH003"]
# Tool scripts use Path at runtime in argparse/signatures
"tools/*.py" = ["TCH003"]

[tool.ruff.lint.isort]
force-single-line = false
combine-as-imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"
markers = [
    "slow: marks tests that require external tools (deselect with '-m \"not slow\"')",
    "container: marks tests that require built Docker images (deselect with '-m \"not container\"')",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = "Bio.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
# structure.py calls extensively into untyped BioPython; no stubs available
module = "autoantibody.structure"
disallow_untyped_calls = false
warn_return_any = false

[tool.hatch.build.targets.wheel]
packages = ["src/autoantibody"]
